% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplyr-package.R
\docType{package}
\name{qplyr-package}
\alias{qplyr}
\alias{qplyr-package}
\title{qplyr: Delayed evaluation with tidyverse verbs}
\description{
Apply tidyverse verbs directly on quosures and eval later.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/qplyr}
  \item Report bugs at \url{https://github.com/moodymudskipper/qplyr/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
